requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
sentence-transformers>=2.7.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
ollama==0.1.7
aiofiles==23.2.1
jinja2==3.1.2
huggingface-hub>=0.20.0
rapidfuzz>=3.0.0
openai-whisper>=20231117
python-multipart>=0.0.6
pypdf>=4.0.0
python-docx>=1.1.0
